Add multiple audio mergers feature
Writed the file of "audio_mixer.h" in order to merger multiple audio into one. which just like the group voice chat.

+++ CameraRtmpSDK/PushSDK/ABSSimpleSession.mm
@@ -15,11 +15,11 @@
 
 @interface ABSSimpleSession()
 {
-    dispatch_queue_t _graphManageQueue;
-    PushSDK::Apple::CameraSource* m_cameraSource;
-    PushSDK::Apple::MicSource* m_micSource;
-    PushSDK::ffmpeg::H264Encode* m_h264Encoder;
-    std::shared_ptr<PushSDK::ffmpeg::AACEncoder> m_aacEncoder;
+  dispatch_queue_t _graphManageQueue;
+  PushSDK::Apple::CameraSource* m_cameraSource;
+  PushSDK::Apple::MicSource* m_micSource;
+  PushSDK::ffmpeg::H264Encode* m_h264Encoder;
+  std::shared_ptr<PushSDK::ffmpeg::AACEncoder> m_aacEncoder;
 }
 @property (nonatomic, assign) int bitrate;
 @property (nonatomic, assign) CGSize videoSize;
@@ -35,21 +35,21 @@
 @implementation ABSSimpleSession
 - (void)dealloc
 {
-    [self endVidoeRecord];
-    [self endAudioRecord];
+  [self endVidoeRecord];
+  [self endAudioRecord];
 
-    Cplus_Release(m_cameraSource);
-    Cplus_Release(m_h264Encoder);
-    Cplus_Release(m_micSource);
+  Cplus_Release(m_cameraSource);
+  Cplus_Release(m_h264Encoder);
+  Cplus_Release(m_micSource);
 }
 
 - (instancetype)init
 {
-    if (self = [super init]) {
+  if (self = [super init]) {
 
 
-    }
-    return self;
+  }
+  return self;
 }
 
 - (instancetype)initWithVideoSize:(CGSize)videoSize
@@ -59,81 +59,81 @@
                       cameraState:(ABSCameraState)cameraState
                      previewFrame:(CGRect)previewFrame
 {
-    if (self = [super init]) {
+  if (self = [super init]) {
 
-        self.videoSize = videoSize;
-        self.fps = fps;
-        self.bitrate = bps;
-        self.useInterfaceOrientation = useInterfaceOrientation;
-        self.cameraState = cameraState;
+    self.videoSize = videoSize;
+    self.fps = fps;
+    self.bitrate = bps;
+    self.useInterfaceOrientation = useInterfaceOrientation;
+    self.cameraState = cameraState;
 
-        _previewView = [[ABSPreviewView alloc] initWithFrame:CGRectMake(0, 0, previewFrame.size.width, previewFrame.size.height)];
+    _previewView = [[ABSPreviewView alloc] initWithFrame:CGRectMake(0, 0, previewFrame.size.width, previewFrame.size.height)];
 
-        _graphManageQueue = dispatch_queue_create("com.PushSDK.session.graph", DISPATCH_QUEUE_SERIAL);
+    _graphManageQueue = dispatch_queue_create("com.PushSDK.session.graph", DISPATCH_QUEUE_SERIAL);
 
-        __weak typeof(self) bSelf = self;
-        //        dispatch_async(_graphManageQueue, ^{
-        [bSelf setupGraph];
-        //        });
-    }
-    return self;
+    __weak typeof(self) bSelf = self;
+    //        dispatch_async(_graphManageQueue, ^{
+    [bSelf setupGraph];
+    //        });
+  }
+  return self;
 }
 
 
 - (instancetype)initWithAudioSampleRate:(float)rate
                            channelCount:(int)count;
 {
-    if (self = [super init]) {
-        self.audioSampleRate = rate > 0 ? rate : (float) 44100.; // 最多只能用 44KHZ，因为 flv 格式不支持 44KHZ 以上。
-        self.audioChannelCount = count > 0 ? count : 2;
+  if (self = [super init]) {
+    self.audioSampleRate = rate > 0 ? rate : (float) 44100.; // 最多只能用 44KHZ，因为 flv 格式不支持 44KHZ 以上。
+    self.audioChannelCount = count > 0 ? count : 2;
 
-        [self setUpMicSource];
-    }
-    return self;
+    [self setUpMicSource];
+  }
+  return self;
 }
 
 - (void)setupGraph
 {
-    m_cameraSource = new PushSDK::Apple::CameraSource();
-    m_cameraSource->setupCamera(self.fps, (_cameraState == ABSCameraStateFront), true, AVCaptureSessionPreset640x480);
-    CALayer* previeLayer = m_cameraSource->captureVideoPreviewLayer();
-    previeLayer.frame = self.previewView.bounds;
-    [self.previewView.layer addSublayer:previeLayer];
+  m_cameraSource = new PushSDK::Apple::CameraSource();
+  m_cameraSource->setupCamera(self.fps, (_cameraState == ABSCameraStateFront), true, AVCaptureSessionPreset640x480);
+  CALayer* previeLayer = m_cameraSource->captureVideoPreviewLayer();
+  previeLayer.frame = self.previewView.bounds;
+  [self.previewView.layer addSublayer:previeLayer];
 }
 
 - (void)setUpMicSource
 {
-    m_micSource = new PushSDK::Apple::MicSource(self.audioSampleRate, self.audioChannelCount);
+  m_micSource = new PushSDK::Apple::MicSource(self.audioSampleRate, self.audioChannelCount);
 }
 
 
 - (void)startVideoRecord
 {
-    m_h264Encoder = new PushSDK::ffmpeg::H264Encode(self.videoSize.width, self.videoSize.height, self.fps, self.bitrate);
-    m_cameraSource->setOutput(m_h264Encoder);
-    m_cameraSource->startRecord();
+  m_h264Encoder = new PushSDK::ffmpeg::H264Encode(static_cast<int>(self.videoSize.width), static_cast<int>(self.videoSize.height), self.fps, self.bitrate);
+  m_cameraSource->setOutput(m_h264Encoder);
+  m_cameraSource->startRecord();
 }
 
 - (void)startAudioRecord
 {
-    NSString *date = nil;
-    NSDateFormatter *formatter = [[NSDateFormatter alloc] init];
-    formatter.dateFormat = @"YYYY-MM-dd hh:mm:ss";
-    date = [formatter stringFromDate:[NSDate date]];
-    NSString* file_path = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, true).lastObject stringByAppendingPathComponent:[NSString stringWithFormat:@"%@.aac", date]];
-    m_aacEncoder = std::shared_ptr<PushSDK::ffmpeg::AACEncoder>(new PushSDK::ffmpeg::AACEncoder((int) self.audioSampleRate, self.audioChannelCount, 96000, [file_path cStringUsingEncoding:NSUTF8StringEncoding]));
-    m_micSource->setOutput(m_aacEncoder);
-    m_micSource->start();
+  NSString *date = nil;
+  NSDateFormatter *formatter = [[NSDateFormatter alloc] init];
+  formatter.dateFormat = @"YYYY-MM-dd hh:mm:ss";
+  date = [formatter stringFromDate:[NSDate date]];
+  NSString* file_path = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, true).lastObject stringByAppendingPathComponent:[NSString stringWithFormat:@"%@.aac", date]];
+  m_aacEncoder = std::shared_ptr<PushSDK::ffmpeg::AACEncoder>(new PushSDK::ffmpeg::AACEncoder((int) self.audioSampleRate, self.audioChannelCount, 96000, [file_path cStringUsingEncoding:NSUTF8StringEncoding]));
+  m_micSource->setOutput(m_aacEncoder);
+  m_micSource->start();
 }
 
 - (void)endAudioRecord
 {
-     m_micSource->stop();
+  m_micSource->stop();
 }
 
 - (void)endVidoeRecord
 {
-     m_cameraSource->stopRecord();
+  m_cameraSource->stopRecord();
 }
